pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = ""
authors = [
    {name = "Kwang Yong Jung",email = "gyjong@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "streamlit (>=1.46.1,<2.0.0)",
    "langchain (>=0.3.26,<0.4.0)",
    "langchain-core (>=0.3.66,<0.4.0)",
    "langchain-community (>=0.3.26,<0.4.0)",
    "langchain-ollama (>=0.3.3,<0.4.0)",
    "langchain-huggingface (>=0.3.0,<0.4.0)",
    "langchain-experimental (>=0.3.4,<0.4.0)",
    "ollama (>=0.5.1,<0.6.0)",
    "streamlit-option-menu (>=0.4.0,<0.5.0)",
    "faiss-cpu (>=1.11.0,<2.0.0)",
    "sentence-transformers (>=4.1.0,<5.0.0)",
    "transformers (>=4.53.0,<5.0.0)",
    "torch (>=2.7.1,<3.0.0)",
    "pypdf (>=5.6.1,<6.0.0)",
    "pdfplumber (>=0.11.7,<0.12.0)",
    "python-docx (>=1.2.0,<2.0.0)",
    "python-pptx (>=1.0.2,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pandas (>=2.3.0,<3.0.0)",
    "numpy (>=2.3.1,<3.0.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "plotly (>=6.2.0,<7.0.0)",
    "scikit-learn (>=1.5.2,<2.0.0)",
    "chromadb (>=1.0.13,<2.0.0)",
    "langchain-milvus (>=0.2.0,<0.3.0)",
    "pymilvus (>=2.5.0,<3.0.0)",
    "tiktoken (>=0.9.0,<0.10.0)",
    "requests (>=2.32.4,<3.0.0)",
    "langchain-text-splitters (>=0.3.8,<0.4.0)",
    "rdflib (>=7.1.4,<8.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "networkx (>=3.5,<4.0)",
    "streamlit-authenticator (>=0.4.2,<0.5.0)",
    "python-louvain (>=0.16,<0.17)",
    "pytest (>=8.4.1,<9.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "flake8 (>=7.3.0,<8.0.0)",
    "mypy (>=1.16.1,<2.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "ipykernel (>=6.29.5,<7.0.0)",
    "googlesearch-python (>=1.2.5,<2.0.0)",
    "beautifulsoup4 (>=4.12.3,<5.0.0)",
    "lxml (>=5.3.0,<6.0.0)",
    "urllib3 (>=2.3.0,<3.0.0)",
    "langgraph (>=0.5.0,<0.6.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
